fix(publish): fix version conflict recognition for github packages
The version conflict recognition logic was not working for github
because it returns a different `code` for the same error compared to
what npmjs.org returns.

I've added two utility functions to check for the error code and
message to make the code more readable in the error handler of the
publish operation.

This way there is a robust detection supporting both npmjs.com and
npm.pkg.github.com registries, not just npmjs.org.

Signed-off-by: Peter Somogyvari <peter.metz@unarin.com>